+++ Source/DREAM3DLib/StatisticsFilters/FindShapes.cpp
@@ -206,7 +206,7 @@ void FindShapes::find_moments()
   {
       grainmoments[i] = 0.0f;
   }
-  for (size_t j = 0; j < totalPoints; j++)
+  for (int64_t j = 0; j < totalPoints; j++)
   {
     u200 = 0;
     u020 = 0;
